requirements.txt
python-nmap
sqlmap
pyfiglet